+++ web/app/themes/brocooly/src/helpers.php
@@ -2,6 +2,6 @@
 /**
  * Theme helper functions and filters
  *
- * @package Brcooly
+ * @package Brocooly
  * @since 1.0.0
  */